Move markdown related file to markdown.py
